grc: move get_state_dir function to remove imports from main
Removes nasty import from main
