+++ grc/gui_qt/Config.py
@@ -3,14 +3,14 @@ from os.path import expanduser, normpath, expandvars, exists
 from collections import OrderedDict
 
 from ..core.Config import Config as CoreConfig
-from ..main import get_config_file_path
+from .. import paths
 from qtpy import QtCore
 
 
 class Config(CoreConfig):
     name = 'GNU Radio Companion'
 
-    gui_prefs_file = os.environ.get('GRC_QT_PREFS_PATH', get_config_file_path('grc_qt.conf'))
+    gui_prefs_file = os.environ.get('GRC_QT_PREFS_PATH', paths.get_config_file_path('grc_qt.conf'))
 
     def __init__(self, install_prefix, *args, **kwargs):
         CoreConfig.__init__(self, *args, **kwargs)
