+++ grc/main.py
@@ -7,15 +7,12 @@
 #
 
 import argparse
-import gettext
-import locale
 import logging
 import logging.handlers
 import os
-import platform
 import sys
-from pathlib import Path
 
+from . import paths
 
 VERSION_AND_DISCLAIMER_TEMPLATE = """\
 GNU Radio Companion (%s) -
@@ -54,7 +51,6 @@ except ImportError as ex:
 
     # Throw the new exception
     raise Exception("Cannot find GNU Radio!") from None
-
 
 # Enable Logging
 # Do this globally so it is available for both run_gtk() and run_qt()
@@ -188,75 +184,9 @@ def run_qt(args, log):
     sys.exit(app.run())
 
 
-def get_config_file_path(config_file: str = "grc.conf") -> str:
-    oldpath = os.path.join(os.path.expanduser("~/.gnuradio"), config_file)
-    try:
-        from gnuradio.gr import paths
-        newpath = os.path.join(paths.userconf(), config_file)
-        if os.path.exists(newpath):
-            return newpath
-        if os.path.exists(oldpath):
-            log.warn(f"Found specification for config path '{newpath}', but file does not exist. " +
-                     f"Old default config file path '{oldpath}' exists; using that. " +
-                     "Please consider moving configuration to new location.")
-            return oldpath
-        # Default to the correct path if both are configured.
-        # neither old, nor new path exist: create new path, return that
-        path_parts = Path(newpath).parts[:-1]
-        pathdir = os.path.join(*path_parts)
-        os.makedirs(pathdir, exist_ok=True)
-        return newpath
-    except ImportError:
-        log.warn("Could not retrieve GNU Radio configuration directory from GNU Radio. Trying defaults.")
-        xdgconf = os.getenv("XDG_CONFIG_HOME", os.path.expanduser("~/.config"))
-        xdgcand = os.path.join(xdgconf, config_file)
-        if os.path.exists(xdgcand):
-            return xdgcand
-        if os.path.exists(oldpath):
-            log.warn(f"Using legacy config path '{oldpath}'. Please consider moving configuration " +
-                     f"files to '{xdgcand}'.")
-            return oldpath
-        # neither old, nor new path exist: create new path, return that
-        path_parts = Path(xdgcand).parts[:-1]
-        pathdir = os.path.join(*path_parts)
-        os.makedirs(pathdir, exist_ok=True)
-        return xdgcand
-
-
-def get_state_directory() -> str:
-    oldpath = os.path.expanduser("~/.grc_gnuradio")
-    try:
-        from gnuradio.gr import paths
-        newpath = paths.persistent()
-        if os.path.exists(newpath):
-            return newpath
-        if os.path.exists(oldpath):
-            log.warn(f"Found specification for persistent state path '{newpath}', but file does not exist. " +
-                     f"Old default persistent state path '{oldpath}' exists; using that. " +
-                     "Please consider moving state to new location.")
-            return oldpath
-        # Default to the correct path if both are configured.
-        # neither old, nor new path exist: create new path, return that
-        os.makedirs(newpath, exist_ok=True)
-        return newpath
-    except (ImportError, NameError):
-        log.warn("Could not retrieve GNU Radio persistent state directory from GNU Radio. Trying defaults.")
-        xdgstate = os.getenv("XDG_STATE_HOME", os.path.expanduser("~/.local/state"))
-        xdgcand = os.path.join(xdgstate, "gnuradio")
-        if os.path.exists(xdgcand):
-            return xdgcand
-        if os.path.exists(oldpath):
-            log.warn(f"Using legacy state path '{oldpath}'. Please consider moving state " +
-                     f"files to '{xdgcand}'.")
-            return oldpath
-        # neither old, nor new path exist: create new path, return that
-        os.makedirs(xdgcand, exist_ok=True)
-        return xdgcand
-
-
 def main():
     grc_version_from_config = ""
-    grc_qt_config_file = get_config_file_path('grc_qt.conf')
+    grc_qt_config_file = paths.get_config_file_path('grc_qt.conf')
     if os.path.isfile(grc_qt_config_file):
         try:
             from qtpy.QtCore import QSettings
@@ -294,7 +224,7 @@ def main():
     log.info("Starting GNU Radio Companion {} (Python {})".format(gr.version(), py_version))
 
     # File logging
-    log_file = os.path.join(get_state_directory(), "grc.log")
+    log_file = os.path.join(paths.get_state_directory(), "grc.log")
     try:
         fileHandler = logging.FileHandler(log_file)
         file_msg_format = '%(asctime)s [%(levelname)s] %(message)s'
